added tasks list in table

+++ ToDoApp/controllers/TaskController.js
@@ -21,23 +21,20 @@ router.post('/add', isAuthenticated, async (req, res) => {
 });
 
 router.get('/:taskId/details', isAuthenticated, async (req, res) => {
-    console.log('Details for Task ID: ', req.params.taskId)
-    const task = await taskService.getTaskById(req.params.taskId);
-    console.log('Task: ', task);
-    //TODO if the task user ID is not the current user redirect to 404
-    // if (task && task.userId !== req.user.id) {
-    //     return res.redirect('/tasks');
-    // }
+    const taskId = req.params.taskId;
+    const task = await taskService.getTaskById(taskId);
+    if (task.userId.toString() !== req.user.id){
+        return res.redirect('/404');
+    }
     res.render('tasks/details', {task});
 });
 
 router.get('/:taskId/edit', isAuthenticated, async (req, res) => {
     const task = await taskService.getTaskById(req.params.taskId);
-    //TODO if the task user ID is not the current user redirect to 404
-    // if (task && task.userId !== req.user.id) {
-    //     return res.redirect('/tasks');
-    // }
     if (!task){
+        return res.redirect('/404');
+    }
+    if (task.userId.toString() !== req.user.id){
         return res.redirect('/404');
     }
     res.render('tasks/edit', {task});
@@ -46,12 +43,11 @@ router.get('/:taskId/edit', isAuthenticated, async (req, res) => {
 router.post('/:taskId/edit', isAuthenticated, async (req, res) => {
     const taskId = req.params.taskId;
     let task = await taskService.getTaskById(taskId);
-    //TODO if the task user ID is not the current user redirect to 404
-    // if (task && task.userId !== req.user.id) {
-    //     return res.redirect('/tasks');
-    // }
     if (!task){
-        return res.redirect('/404');
+        return res.status(404).end();
+    }
+    if (task.userId.toString() !== req.user.id){
+        return res.status(404).end();
     }
     const changedTaskInputValues = req.body;
     if (task.title !== changedTaskInputValues.title){
@@ -72,20 +68,21 @@ router.post('/:taskId/edit', isAuthenticated, async (req, res) => {
 
 router.get('/:taskId/delete', isAuthenticated, async (req, res) => {
     const task = await taskService.getTaskById(req.params.taskId);
-    //TODO if the task user ID is not the current user redirect to 404
-    // if (task && task.userId !== req.user.id) {
-    //     return res.redirect('/tasks');
-    // }
+    if (task.userId.toString() !== req.user.id){
+        return res.redirect('/404');
+    }
     res.render('tasks/delete', {task});
 });
 
 router.post('/:taskId/delete', isAuthenticated, async (req, res) => {
     const taskId = req.params.taskId;
     const task = await taskService.getTaskById(taskId);
-    //TODO if the task user ID is not the current user redirect to 404
-    // if (task && task.userId !== req.user.id) {
-    //     return res.redirect('/tasks');
-    // }
+    if (!task){
+        return res.status(404).end();
+    }
+    if (task.userId.toString() !== req.user.id){
+        return res.status(404).end();
+    }
     const message = await taskService.delete(taskId);
     console.log(message);
     res.redirect('/tasks');
